When the ready event fail

+++ storage/commands/destroy.js
@@ -2,11 +2,19 @@ const Discord = require("discord.js");
 const config = require("../config.json");
 const admin = ["210392675478667269"]
 
-module.exports.run = async (bot, message, dataSpecialChannel)=>{
-    message.delete();
-    message.channel.send("Arrêt du bot")
-        .then(msg => {msg.delete({ timeout: 1500 })});
-    bot.user.setActivity("Arrêt en cours", {type: "WATCHING"});
+module.exports.run = async (bot, message, dataSpecialChannel) => {
+    if (message != null) {
+        message.delete();
+        message.channel.send("Arrêt du bot")
+            .then(msg => {
+                msg.delete({
+                    timeout: 1500
+                })
+            });
+    }
+    bot.user.setActivity("Arrêt en cours", {
+        type: "WATCHING"
+    });
     await bot.basicFunctions.get("wait").run(2500);
     bot.destroy();
     bot.enventIndex.get("cronTab").stop(bot);
@@ -20,7 +28,7 @@ module.exports.help = {
 module.exports.manuel = {
     man: "man permet d'afficher le manuel d'une fonction\n" +
         "Pour ce faire il faut écrire dans le terminal:\n" +
-        "`"+config.prefix+" man [nom de la commande]`"
+        "`" + config.prefix + " man [nom de la commande]`"
 };
 
 module.exports.manuelview = {
